+++ src/hal/macos/network_interface_macos.c
@@ -1,0 +1,28 @@
+/**
+ * File:   network_interface_macos.c
+ * Author: AWTK Develop Team
+ * Brief: network_interface for macos
+ *
+ * Copyright (c) 2019 - 2021  Guangzhou ZHIYUAN Electronics Co.,Ltd.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * License file for more details.
+ *
+ */
+
+/**
+ * History:
+ * ================================================================
+ * 2021-02-3 pengguowen<pengguowen@zlg.cn> created
+ *
+ */
+
+#include "hal/network_interface.h"
+
+network_interface_t* network_interface_create(const char* interface,
+                                              network_interface_type_t type) {
+  /*TODO*/
+  return NULL;
+}