+++ tools/idl_gen/tkc.js
@@ -13,6 +13,7 @@ let sourcesPath = [
     path.normalize(path.join(__dirname, '../../src/ubjson')), 
     path.normalize(path.join(__dirname, '../../src/csv')), 
     path.normalize(path.join(__dirname, '../../src/streams')), 
+    path.normalize(path.join(__dirname, '../../src/hal')), 
     path.normalize(path.join(__dirname, '../../src/conf_io'))].join(';');
 
 IDLGen.gen(sourcesPath, outputIDL)